Convert openstack groovy script to a helper function
