+++ .rubocop.yml
@@ -48,6 +48,7 @@ Style/Documentation:
     - 'spec/**/*'
     - 'test/**/*'
     - 'libraries/docker.rb'
+    - 'libraries/openstack.rb'
     - 'libraries/github_oauth.rb'
     - 'recipes/powerci.rb'
     - 'recipes/ibmz_ci.rb'
