+++ recipes/powerci.rb
@@ -19,6 +19,7 @@
 
 class ::Chef::Recipe
   include OSLDocker::Helper
+  include OSLOpenStack::Helper
   include OSLGithubOauth::Helper
 end
 
@@ -178,6 +179,8 @@ docker_cloud =
     nil
   )
 
+openstack_cloud = add_openstack_cloud
+
 github_oauth =
   add_github_oauth(
     client_id,
@@ -186,73 +189,12 @@ github_oauth =
     normal_users
   )
 
-file '/tmp/docker.groovy' do
-  content docker_cloud
-end
-
 jenkins_script 'Add Docker Cloud' do
   command docker_cloud
 end
 
 jenkins_script 'Add OpenStack Cloud' do
-  command <<-EOH.gsub(/^ {4}/, '')
-    import jenkins.plugins.openstack.compute.*
-		import jenkins.model.*
-    import hudson.model.*;
-
-    def instance = Jenkins.getInstance()
-
-    JCloudsSlaveTemplate template = new JCloudsSlaveTemplate(
-      "template",
-      "label",
-      new SlaveOptions(
-          "img",
-          "hw",
-          "nw",
-          "ud",
-          1,
-          "public",
-          "sg",
-          "az",
-          2,
-          "kp",
-          3,
-          "jvmo",
-          "fsRoot",
-          "cid",
-          JCloudsCloud.SlaveType.JNLP,
-          4
-      )
-    );
-    JCloudsCloud cloud = new JCloudsCloud(
-      "osl-openstack", // name for the openstack cloud
-      "identity", // format: TENANT_NAME:USER_NAME
-      "credential", //password for the username
-      "endPointUrl", //OpenStack Auth URL
-      "zone",
-      new SlaveOptions(
-             "IMG", //Image
-             "HW",
-             "NW",
-             "UD",
-             6,
-             null,
-             "SG",
-             "AZ",
-             7,
-             "KP",
-             8,
-             "JVMO",
-             "FSrOOT",
-             "CID",
-             JCloudsCloud.SlaveType.SSH,
-             9
-      ),
-      Arrays.asList(template)
-    );
-
-    instance.clouds.add(cloud);
-  EOH
+  command openstack_cloud
 end
 
 jenkins_script 'Add GitHub OAuth config' do
